+++ Tulpep.PayULibrary/Models/Response/Response_RecurringPayments/Subscription/Creation/NewPlan/RootPayUSubscriptionCreationNewPlanResponse.cs
@@ -7,7 +7,7 @@
         public Response_Subscription_Creation_NewPlan_Customer customer { get; set; }
         public string quantity { get; set; }
         public string installments { get; set; }
-        public long currentPeriodStart { get; set; }
-        public long currentPeriodEnd { get; set; }
+        public double currentPeriodStart { get; set; }
+        public double currentPeriodEnd { get; set; }
     }
 }
